- Added OAUTH2 with JWT authentication and authorization for the database API - Now the passwords are stored as hashes created with Bcrypt hashing algorithm - Created a database API proxy class that:     - Encapsulates the communication logic with the database API     - Acquires/renews JWT from the database using the given credentials - Secured most of the database API endpoints - Created automated tests for identity management function and the API proxy

+++ src/persistence/routes/SongRoutes.py
@@ -1,50 +1,75 @@
+from http import HTTPStatus
 from typing import Any
 
 from classy_fastapi import get, patch, post, delete
-from fastapi import Body
-from starlette import status
-from starlette.responses import Response, JSONResponse
+from fastapi import Body, Depends, HTTPException
+from starlette.responses import JSONResponse
 
 import config.database_api as api_paths
+from config.user_types import USER, MICROSERVICE
 from src.helpers.ModelUtils import orm_to_dict
 from src.model.orm.Song import Song
 from src.persistence.routes.abstract_classes.AbstractRoutable import AbstractRoutable
+from src.persistence.routes.abstract_classes.AbstractSecuredRoutable import AbstractSecuredRoutable
 
 
-class SongRoutes(AbstractRoutable):
+class SongRoutes(AbstractSecuredRoutable):
     @post(api_paths.SONGS_PATH)
-    def post_song(self, body: dict[str, Any] = Body()):
+    def post_song(self, body: dict[str, Any] = Body(),
+                  token: str = Depends(AbstractRoutable.OAUTH2_SCHEME)):
+        self._assert_has_user_type(token, MICROSERVICE)
+
         session = self._create_session()
         song = Song(**body)
         session.add(song)
         session.flush()
         session.commit()
-        return JSONResponse(orm_to_dict(song), status_code=status.HTTP_201_CREATED)
+        return JSONResponse(orm_to_dict(song), status_code=HTTPStatus.CREATED)
 
     @get(api_paths.SONG_BY_ID_PATH)
-    def get_song_by_id(self, song_id: int):
+    def get_song_by_id(self, song_id: int, token: str = Depends(AbstractRoutable.OAUTH2_SCHEME)):
+        payload = self._assert_has_user_type_in(token, [USER, MICROSERVICE])
+
         session = self._create_session()
         song = session.get(Song, song_id)
+
         if song is None:
-            return Response(status_code=status.HTTP_404_NOT_FOUND)
+            raise HTTPException(HTTPStatus.NOT_FOUND)
+
+        if payload['user_type_id'] == USER and song.user_id != payload['user_id']:
+            raise HTTPException(HTTPStatus.FORBIDDEN)
+
         return song
 
     @patch(api_paths.SONG_BY_ID_PATH)
-    def patch_song_by_id(self, song_id: int, body: dict[str, Any] = Body()):
+    def patch_song_by_id(self, song_id: int, body: dict[str, Any] = Body(),
+                         token: str = Depends(AbstractRoutable.OAUTH2_SCHEME)):
+        payload = self._assert_has_user_type_in(token, [MICROSERVICE, USER])
+
         session = self._create_session()
         query = session.query(Song).filter_by(song_id=song_id)
         if query.count() == 0:
-            return Response(status_code=status.HTTP_404_NOT_FOUND)
+            raise HTTPException(HTTPStatus.NOT_FOUND)
+
+        song = query.first()
+        if payload['user_type_id'] == USER and payload['user_id'] != song.user_id:
+            raise HTTPException(HTTPStatus.FORBIDDEN)
+
         query.update(body)
         session.commit()
-        return Response(status_code=status.HTTP_200_OK)
 
     @delete(api_paths.SONG_BY_ID_PATH)
-    def delete_song(self, song_id: int):
+    def delete_song(self, song_id: int, token: str = Depends(AbstractRoutable.OAUTH2_SCHEME)):
+        payload = self._assert_has_user_type(token, USER)
+
         session = self._create_session()
-        song = session.get(Song, song_id)
+        song: Song = session.get(Song, song_id)
+
         if song is None:
-            return Response(status_code=status.HTTP_404_NOT_FOUND)
+            raise HTTPException(HTTPStatus.NOT_FOUND)
+
+        if song.user_id != payload['user_id']:
+            raise HTTPException(HTTPStatus.FORBIDDEN)
+
         session.delete(song)
         session.commit()
-        return Response(status_code=status.HTTP_200_OK)
